+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/RectangleSelection/dataNormalizers/RatioNormalizer.js
@@ -1,0 +1,27 @@
+// @flow
+import type {DataNormalizer, SelectionData} from '../types';
+
+export default class RatioNormalizer implements DataNormalizer {
+    width: number;
+    height: number;
+
+    constructor(width: number, height: number) {
+        this.width = width;
+        this.height = height;
+    }
+
+    normalize(data: SelectionData): SelectionData {
+        let height = data.height;
+        let width = data.width;
+        let calculatedHeight = width * (this.height / this.width);
+        let calculatedWidth = height * (this.width / this.height);
+
+        if (calculatedWidth <= width) {
+            width = calculatedWidth;
+        } else {
+            height = calculatedHeight;
+        }
+
+        return {...data, width, height};
+    }
+}