+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/RectangleSelection/dataNormalizers/RoundingNormalizer.js
@@ -1,0 +1,13 @@
+// @flow
+import type {DataNormalizer, SelectionData} from '../types';
+
+export default class RoundingNormalizer implements DataNormalizer {
+    normalize(data: SelectionData): SelectionData {
+        return {
+            width: Math.round(data.width),
+            height: Math.round(data.height),
+            top: Math.round(data.top),
+            left: Math.round(data.left),
+        };
+    }
+}