+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/RectangleSelection/index.js
@@ -1,0 +1,4 @@
+// @flow
+import RectangleSelection from './RectangleSelection';
+
+export default RectangleSelection;