+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/RectangleSelection/types.js
@@ -1,0 +1,18 @@
+// @flow
+export interface DataNormalizer {
+    normalize(data: SelectionData): SelectionData,
+}
+
+export type SelectionData = {
+    left: number,
+    top: number,
+    width: number,
+    height: number,
+};
+
+export type RectangleChange = {
+    top: number,
+    left: number,
+    width: number,
+    height: number,
+};